Mechanism to avoid interaction between prefs used during integration tests and current user prefs. Now : when running integration tests : a new Instance of the applicationDataDirectory is used (a temp directory) and so new preferences are used for each test. (applicationDataDirectory is created during test setup and clean up during tearDown). (similar mechanism for FlexoResourceCenter was done in a previous commit, but not really explained in commit message)
I had to remove the old feature retrieving the "oldPrefs" from the "legacy" prefs location. Impact of this removal is really minor since the default pref location was changed a few years ago and so it may have a really small impact only on users who didn't use flexo during the last 3/4 years.